% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycomb.R
\name{tidy_comb}
\alias{tidy_comb}
\title{tidy_comb}
\usage{
tidy_comb(vec)
}
\arguments{
\item{vec}{a list of elements to combine}
}
\value{
a tibble with all possible combination of a list
}
\description{
tidy_comb
}
\examples{
tidy_comb(vec = LETTERS[1:5])
}
