% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidystringdist.R
\name{tidy_stringdist}
\alias{tidy_stringdist}
\title{Tidy stringdist calculation}
\usage{
tidy_stringdist(list, method = c("osa", "lv", "dl", "hamming", "lcs", "qgram",
  "cosine", "jaccard", "jw", "soundex"))
}
\arguments{
\item{list}{a list of strings to compare}

\item{method}{one of the methods implemented in the stringdist package — "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex". See \code{\link[stringdist]{stringdist-metrics}}}
}
\value{
a tibble with string distance
}
\description{
Tidy stringdist calculation
}
\examples{
proust <- c("Albertine", "Françoise", "Gilberte", "Odette", "Charles")
tidy_stringdist(proust,method= "osa")
}
