% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycomb.R
\name{tidy_comb}
\alias{tidy_comb}
\alias{tidy_comb.data.frame}
\alias{tidy_comb.default}
\title{Tidy combine}
\usage{
tidy_comb(data, base, ...)

\method{tidy_comb}{data.frame}(data, base, ...)

\method{tidy_comb}{default}(data, base, ...)
}
\arguments{
\item{data}{data object containing the list of words, either a list or a data.frame}

\item{base}{the base word to compare with all the words}

\item{...}{if data is a data.frame, the col where the words to combine are}
}
\value{
a tibble with all possible combination of elements from a list
}
\description{
Get all combinations from a dataframe column or from a list
}
\examples{
tidy_comb(iris, "this", Species)
tidy_comb(state.name, "Paris")
}
